+++ src/vm/vm.h
@@ -113,7 +113,7 @@ public:
 	unsigned long Uptime();
 	unsigned long GetElapsedTime(int instance);
 	void ListPlayingSounds();
-	int GetGlobalStats(int &meshes, int &textures, int &actions, int &sounds, int &objects, int &walls, int &polygons);
+	unsigned long GetGlobalStats(unsigned long &meshes, unsigned long &textures, unsigned long &actions, unsigned long &sounds, unsigned long &objects, unsigned long &walls, unsigned long &polygons);
 
 	bool Shutdown;
 	bool ConcurrentLoads; //set to true for buildings to be loaded while another sim is active and rendering
